requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.11.0
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.18.9
botocore==1.21.9
cachetools==5.3.1
celery==5.2.7
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.3.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.1
decorator==5.1.0
Delorean==1.0.0
Django==3.2
django-admin-rangefilter==0.8.1
django-axes==6.1.0
django-captcha==1.0.0
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-ckeditor==6.5.1
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.1
django-js-asset==2.0.0
django-phrase==1.0.8
django-recaptcha==3.0.0
django-related-admin==1.0.1
django-storages==1.11.1
django-timezone-field==5.0
django-widget-tweaks==1.4.8
django-xhtml2pdf==0.0.4
frozenlist==1.4.0
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
gunicorn==20.0.4
html5lib==1.1
httplib2==0.22.0
hubspot==0.1.14.dev0
hubspot-api-client==7.3.0
humanize==4.5.0
idna==2.10
ipaddress==1.0.16
ipdb==0.13.9
ipython==7.28.0
jedi==0.18.2
jmespath==0.10.0
kombu==5.2.4
line-profiler==4.0.3
matplotlib-inline==0.1.3
multidict==6.0.4
mysqlclient==2.0.3
numpy==1.22.1
oauthlib==3.2.2
openai==0.28.1
pandas==1.4.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.36
protobuf==4.24.4
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycurl==7.45.1
pygeoip==0.3.2
Pygments==2.14.0
pyparsing==3.0.9
PyPDF2==1.26.0
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
reportlab==3.6.1
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.5.0
simplejson==3.18.1
six==1.16.0
sqlparse==0.4.1
stripe==2.56.0
texttable==1.6.4
toml==0.10.2
tqdm==4.66.1
traitlets==5.9.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.6
vine==5.0.0
watchtower==1.0.6
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==0.59.0
xhtml2pdf==0.2.5
xlwt==1.3.0
yarl==1.9.2
